chore: refactor packet data v2 to use type with suffix in v2 proto package
