+++ modules/apps/transfer/internal/convert/convert_test.go
@@ -8,7 +8,6 @@ import (
 	errorsmod "cosmossdk.io/errors"
 
 	"github.com/cosmos/ibc-go/v8/modules/apps/transfer/types"
-	v3types "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types/v3"
 )
 
 func TestConvertPacketV1ToPacketV3(t *testing.T) {
@@ -20,14 +19,14 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 	testCases := []struct {
 		name     string
 		v1Data   types.FungibleTokenPacketData
-		v3Data   v3types.FungibleTokenPacketData
+		v2Data   types.FungibleTokenPacketDataV2
 		expPanic error
 	}{
 		{
 			"success",
 			types.NewFungibleTokenPacketData("transfer/channel-0/atom", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom",
 						Amount: "1000",
@@ -39,8 +38,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success with empty trace",
 			types.NewFungibleTokenPacketData("atom", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom",
 						Amount: "1000",
@@ -52,8 +51,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success: base denom with '/'",
 			types.NewFungibleTokenPacketData("transfer/channel-0/atom/withslash", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom/withslash",
 						Amount: "1000",
@@ -65,8 +64,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success: base denom with '/' at the end",
 			types.NewFungibleTokenPacketData("transfer/channel-0/atom/", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom/",
 						Amount: "1000",
@@ -78,8 +77,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success: longer trace base denom with '/'",
 			types.NewFungibleTokenPacketData("transfer/channel-0/transfer/channel-1/atom/pool", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom/pool",
 						Amount: "1000",
@@ -91,8 +90,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success: longer trace with non transfer port",
 			types.NewFungibleTokenPacketData("transfer/channel-0/transfer/channel-1/transfer-custom/channel-2/atom", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom",
 						Amount: "1000",
@@ -104,8 +103,8 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"success: base denom with slash, trace with non transfer port",
 			types.NewFungibleTokenPacketData("transfer/channel-0/transfer/channel-1/transfer-custom/channel-2/atom/pool", "1000", sender, receiver, ""),
-			v3types.NewFungibleTokenPacketData(
-				[]*v3types.Token{
+			types.NewFungibleTokenPacketDataV2(
+				[]*types.Token{
 					{
 						Denom:  "atom/pool",
 						Amount: "1000",
@@ -117,7 +116,7 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 		{
 			"failure: panics with empty denom",
 			types.NewFungibleTokenPacketData("", "1000", sender, receiver, ""),
-			v3types.FungibleTokenPacketData{},
+			types.FungibleTokenPacketDataV2{},
 			errorsmod.Wrap(types.ErrInvalidDenomForTransfer, "base denomination cannot be blank"),
 		},
 	}
@@ -125,11 +124,11 @@ func TestConvertPacketV1ToPacketV3(t *testing.T) {
 	for _, tc := range testCases {
 		expPass := tc.expPanic == nil
 		if expPass {
-			v3Data := PacketDataV1ToV3(tc.v1Data)
-			require.Equal(t, tc.v3Data, v3Data, "test case: %s", tc.name)
+			v3Data := PacketDataV1ToV2(tc.v1Data)
+			require.Equal(t, tc.v2Data, v3Data, "test case: %s", tc.name)
 		} else {
 			require.PanicsWithError(t, tc.expPanic.Error(), func() {
-				PacketDataV1ToV3(tc.v1Data)
+				PacketDataV1ToV2(tc.v1Data)
 			}, "test case: %s", tc.name)
 		}
 	}
