+++ proto/ibc/applications/transfer/v2/packet.proto
@@ -19,3 +19,27 @@ message FungibleTokenPacketData {
   // optional memo
   string memo = 5;
 }
+
+// FungibleTokenPacketDataV2 defines a struct for the packet payload
+// See FungibleTokenPacketDataV2 spec:
+// https://github.com/cosmos/ibc/tree/master/spec/app/ics-020-fungible-token-transfer#data-structures
+message FungibleTokenPacketDataV2 {
+  // the tokens to be transferred
+  repeated Token tokens = 1;
+  // the sender address
+  string sender = 2;
+  // the recipient address on the destination chain
+  string receiver = 3;
+  // optional memo
+  string memo = 4;
+}
+
+// Token defines a struct which represents a token to be transferred.
+message Token {
+  // the base token denomination to be transferred
+  string denom = 1;
+  // the token amount to be transferred
+  string amount = 2;
+  // the trace of the token
+  repeated string trace = 3;
+}
