--- proto/ibc/applications/transfer/v3/packet.proto
@@ -1,29 +1,0 @@
-syntax = "proto3";
-
-package ibc.applications.transfer.v3;
-
-option go_package = "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types/v3";
-
-// FungibleTokenPacketData defines a struct for the packet payload
-// See FungibleTokenPacketData spec:
-// https://github.com/cosmos/ibc/tree/master/spec/app/ics-020-fungible-token-transfer#data-structures
-message FungibleTokenPacketData {
-  // the tokens to be transferred
-  repeated Token tokens = 1;
-  // the sender address
-  string sender = 2;
-  // the recipient address on the destination chain
-  string receiver = 3;
-  // optional memo
-  string memo = 4;
-}
-
-// Token defines a struct which represents a token to be transferred.
-message Token {
-  // the base token denomination to be transferred
-  string denom = 1;
-  // the token amount to be transferred
-  string amount = 2;
-  // the trace of the token
-  repeated string trace = 3;
-}
